+++ src/canbus/can_egs52.h
@@ -87,7 +87,7 @@ class Egs52Can: public EgsBaseCan {
         // Sets gearbox is OK
         void set_gearbox_ok(bool is_ok) override;
         // Sets torque request toggle
-        void set_torque_request(TorqueRequest request, int16_t amount_nm) override;
+        void set_torque_request(TorqueRequest request, float amount_nm) override;
         // Sets the status of system error check
         void set_error_check_status(SystemStatusCheck ssc) override;
         // Sets torque loss of torque converter
